requirements.txt
annotated-types==0.7.0
arrow==1.3.0
blinker==1.9.0
branca==0.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fitparse==1.2.0
Flask==3.1.0
flexcache==0.3
flexparser==0.4
folium==0.19.5
gpxpy==1.6.2
greenlet==3.2.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
msgpack==1.1.0
numpy==2.2.5
Pint==0.24.4
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
pyngrok==7.2.5
pynvim==0.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
stravalib==2.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
xyzservices==2025.1.0
